+++ source/launcher.c
@@ -26,12 +26,12 @@ Result launcher_fini(){
 int prepare_to_launch(char* dirname){
 	Handle handle;
 	char* savespath = "/TownManager/Saves/";
-	char* dirpath;
+	char* townpath;
 	char* filepath;
 	char* buf;
 	u32 written;
 	u64 size;
-	file_t files;
+	file_t sdmc_files, game_files;
 	int i;
 	int error;
 	u64 securein;
@@ -39,21 +39,37 @@ int prepare_to_launch(char* dirname){
 	Result ret;
 
 	gfx_displaymessage("Preparing to launch game...");
-	dirpath = calloc(strlen(savespath)+strlen(dirname)+1+1, 1);
-	sprintf(dirpath, "%s%s/", savespath, dirname);
-	files = get_files(sdmc_arch, dirpath);
+	townpath = calloc(strlen(savespath)+strlen(dirname)+1+1, 1);
+	sprintf(townpath, "%s%s/", savespath, dirname);
+	sdmc_files = get_files(sdmc_arch, townpath);
+	game_files = get_files(game_arch, "/");
+
+	//delete files off of game cart if num of files in townpath > 0
+	if(sdmc_files.numfiles > 0){
+		for(i = 0; i < game_files.numfiles; i++){
+			filepath = calloc(1+strlen(game_files.files[i])+1, 1);
+			sprintf(filepath, "/%s", game_files.files[i]);
+			gfx_displaymessage("Deleting %s in game archive...", game_files.files[i]);
+			ret = FSUSER_DeleteFile(game_arch, fsMakePath(PATH_ASCII, filepath));
+			if(ret){
+				gfx_error(ret, __FILENAME__, __LINE__);
+				return -1;
+			}
+		}
+	}
 
 	//copy local files to cartridge
-	for(i = 0; i < files.numfiles; i++){
-		buf = file_to_buffer(sdmc_arch, dirpath, files.files[i]);
+	for(i = 0; i < sdmc_files.numfiles; i++){
+		buf = file_to_buffer(sdmc_arch, townpath, sdmc_files.files[i]);
 		if(buf == NULL){
 			gfx_waitmessage("Buffer is empty!");
 			return -1;
 		}
-		filepath = calloc(strlen(dirpath)+strlen(files.files[i])+1, 1);
-		sprintf(filepath, "%s%s", dirpath, files.files[i]);
+		filepath = calloc(strlen(townpath)+strlen(sdmc_files.files[i])+1, 1);
+		sprintf(filepath, "%s%s", townpath, sdmc_files.files[i]);
 		size = filesize_to_u64(sdmc_arch, filepath);
-		error = buffer_to_file(game_arch, buf, size, "/", files.files[i]);
+		gfx_displaymessage("Writing %s to game archive...", sdmc_files.files[i]);
+		error = buffer_to_file(game_arch, buf, size, "/", sdmc_files.files[i]);
 		if(error == -1){
 			gfx_waitmessage("buffer_to_file failed!");
 			return -1;
@@ -69,7 +85,7 @@ int prepare_to_launch(char* dirname){
 	}
 	if(is3dsx)
 		fs_init();
-	free(dirpath);
+	free(townpath);
 	//Write dirname as the sole contents of the /TownManager/tm.conf file
 	if((ret = FSUSER_OpenFile(&handle, sdmc_arch, fsMakePath(PATH_ASCII, "/TownManager/tm.conf"), FS_OPEN_WRITE, 0))){
 		gfx_error(ret, __FILENAME__, __LINE__);